refactor: update to vlink_token_service and vlink_council v5 contracts across utility files
