+++ aleo/utils/hash_packet.ts
@@ -1,6 +1,6 @@
 import { js2leo as js2leoCommon } from '@doko-js/core';
 import { leo2js as leo2jsCommon } from '@doko-js/core';
-import { InPacket } from '../artifacts/js/types/vlink_token_bridge_v4';
+import { InPacket } from '../artifacts/js/types/vlink_token_bridge_v5';
 import { evm2AleoArr } from './ethAddress';
 import { aleoChainId, ethTsContractAddr } from './constants';
