+++ aleo/utils/sign.ts
@@ -3,8 +3,8 @@ import { js2leo as js2leoCommon } from '@doko-js/core';
 import { sign, sign_verify } from "aleo-signer"
 
 import { hashStruct } from "./hash";
-import { InPacket, InPacketWithScreening } from '../artifacts/js/types/vlink_token_bridge_v4';
-import { getInPacketLeo, getInPacketWithScreeningLeo } from '../artifacts/js/js2leo/vlink_token_bridge_v4';
+import { InPacket, InPacketWithScreening } from '../artifacts/js/types/vlink_token_bridge_v5';
+import { getInPacketLeo, getInPacketWithScreeningLeo } from '../artifacts/js/js2leo/vlink_token_bridge_v5';
 
 export const signPacket = (packet: InPacket, screening_passed: boolean, privateKey: string) => {
