+++ aleo/utils/voters.ts
@@ -1,14 +1,14 @@
 
-import { Vlink_council_v4Contract } from "../artifacts/js/vlink_council_v4";
+import { Vlink_council_v5Contract } from "../artifacts/js/vlink_council_v5";
 
-import { getProposalVoteLeo } from "../artifacts/js/js2leo/vlink_council_v4";
-import { ProposalVote, ProposalVoterKey } from "../artifacts/js/types/vlink_council_v4";
+import { getProposalVoteLeo } from "../artifacts/js/js2leo/vlink_council_v5";
+import { ProposalVote, ProposalVoterKey } from "../artifacts/js/types/vlink_council_v5";
 import { ALEO_ZERO_ADDRESS } from "./testdata.data";
 import { hashStruct } from "./hash";
 import { ExecutionMode } from "@doko-js/core";
 
 const mode = ExecutionMode.SnarkExecute;
-const council = new Vlink_council_v4Contract({ mode });
+const council = new Vlink_council_v5Contract({ mode });
 
 const getVoteKeys = (proposalHash: bigint, voters: string[]): bigint[] => {
   const voteKeys = []
